# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAGE – Smart Adaptive Guidance Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiError {
  error: string;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  title: string;
  content: string;
}

export interface NoteUpdate {
  title?: string;
  content?: string;
}

export interface Todo {
  id: number;
  title: string;
  completed: boolean;
  priority: string;
  createdAt: string;
}

export interface TodoInput {
  title: string;
  priority: string;
}

export interface TodoUpdate {
  title?: string;
  completed?: boolean;
  priority?: string;
}

export interface Stats {
  totalConversations: number;
  totalMessages: number;
  totalNotes: number;
  totalTodos: number;
  completedTodos: number;
}

export interface ApiError {
  error: string;
}