# lib/api-zod/src/generated/types/geminiMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAGE – Smart Adaptive Guidance Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiMessageInput {
  content: string;
}