# lib/api-zod/src/generated/types/noteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAGE – Smart Adaptive Guidance Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface NoteInput {
  title: string;
  content: string;
}